LittleFS driver working and tested
Completed the implementation of the driver, tested using a ramdisk
implementation provided by the lfs library itself.

NOTE: Support of the ramdisk should be removed, it was just added to
test the driver. If we wanted to use a ram block device, then one should
implement it is a Miosix driver